requirements.txt
typer
loguru
python-dotenv